requirements.txt
langchain==0.1.12
langchain-openai==0.0.8
langgraph==0.0.27
python-dotenv==1.0.1
rich==13.7.1
typer==0.9.0 